Nkilani-by-Teiyan-main/onlineshop/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.4
django-daraja==1.3.0
idna==3.4
Pillow==10.0.0
pycparser==2.21
pyngrok==7.0.0
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
subprocess.run==0.0.8
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
